services/inventory_service/requirements.txt
Flask
faker
paramiko
cassandra-driver
psycopg2-binary
kafka-python

